% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.band}
\alias{ec.band}
\title{Band}
\usage{
ec.band(df = NULL, lower = NULL, upper = NULL, ...)
}
\arguments{
\item{df}{A data.frame with lower and upper numerical columns.}

\item{lower}{The column name of band's lower boundary, a string.}

\item{upper}{The column name of band's upper boundary, a string.}

\item{...}{More parameters for \href{https://echarts.apache.org/en/option.html#series-line.type}{serie}}
}
\value{
One serie list
}
\description{
Add a new 'custom' serie with coordinates of a polygon.
}
\details{
The coordinates of the two boundaries are merged into a polygon and displayed as one.
}
