% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ec.clmn}
\alias{ec.clmn}
\title{Data column format}
\usage{
ec.clmn(col = NULL, ..., scale = 1)
}
\arguments{
\item{col}{Can contain one of several types of values:\cr
\tab NULL(default) for charts with single values like tree, pie.\cr
\tab a single column index(number) or column name(quoted string) \cr
\tab a \link[base]{sprintf} string template for multiple columns\cr
\tab 'json' to display tooltip with all available values to choose from\cr
\tab 'log' to write all values in the JS console (F12) for debugging.\cr
\tab a string containing a JS function starting with \emph{'function('} or \emph{'(x) =>'}.\cr}

\item{...}{Comma separated column indexes or names, only when \emph{col} is \emph{sprintf}. This allows formatting of multiple columns, as for a tooltip.\cr}

\item{scale}{A positive number, multiplier for numeric columns. When scale is 0, all numeric values are rounded.}
}
\value{
A JavaScript code string (usually a function) marked as executable, see \link[htmlwidgets]{JS}.
}
\description{
Helper function to display/format data column(s) by index or name
}
\details{
This function is useful for attributes like formatter, color, symbolSize, label.\cr
Column indexes are counted in R and start with 1.\cr
Omit \emph{col} or use index -1 for single values in tree/pie charts, \emph{axisLabel.formatter} or \emph{valueFormatter}. See \link{ec.data} dendrogram example.\cr
Column indexes are decimals for combo charts with multiple series, see \link{ecr.band} example. The whole number part is the serie index, the decimal part is the column index inside.\cr
\emph{col} as sprintf has the same placeholder \emph{\%@} for both column indexes or column names.\cr
\emph{col} as sprintf can contain double quotes, but not single or backquotes.\cr
Template placeholders with formatting:\cr
\itemize{
\item \emph{\%@} will display column value as-is.\cr
\item \emph{\%L@} will display a number in locale format, like '12,345.09'.\cr
\item \emph{\%LR@} rounded number in locale format, like '12,345'.\cr
\item \emph{\%R@} rounded number, like '12345'.\cr
\item \emph{\%R2@} rounded number, two digits after decimal point.\cr
\item \emph{\%M@} marker in series' color.\cr
For \emph{trigger='axis'} (multiple series) one can use decimal column indexes.\cr
See definition above and example below.
}
}
\examples{
library(dplyr)
tmp <- data.frame(Species = as.vector(unique(iris$Species)),
                  emoji = c('A','B','C'))
df <- iris |> inner_join(tmp)      # add 6th column emoji
df |> group_by(Species) |> ec.init(
  series.param= list(label= list(show= TRUE, formatter= ec.clmn('emoji'))),
  tooltip= list(formatter=
    # with sprintf template + multiple column indexes
    ec.clmn('\%M@ species <b>\%@</b><br>s.len <b>\%@</b><br>s.wid <b>\%@</b>', 5,1,2))
)

# tooltip decimal indexes work with full data sets (no missing/partial data)
ChickWeight |> mutate(Chick=as.numeric(Chick)) |> filter(Chick>47) |> group_by(Chick) |>
ec.init(
  tooltip= list(trigger='axis', 
                formatter= ec.clmn("48: \%@<br>49: \%@<br>50: \%@", 1.1, 2.1, 3.1)),
  xAxis= list(type='category'), legend= list(formatter= 'Ch.{name}'),
  series.param= list(type='line', encode= list(x='Time', y='weight')),
)

mtcars |> ec.init(
  series.param= list(encode= list(y='wt'), label= list(show=TRUE, formatter=ec.clmn('\%R2@','wt'))), 
  tooltip= list(valueFormatter= ec.clmn(scale=0) )
)
}
