% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.data}
\alias{ec.data}
\title{Get an EchartsJS dataset from a data.frame}
\usage{
ec.data(df, series = FALSE)
}
\arguments{
\item{df}{Chart data in data.frame format, required}

\item{series}{If FALSE, data is prepared for \href{https://echarts.apache.org/en/option.html#dataset.source}{dataset} (default),\cr
if TRUE, data is for \href{https://echarts.apache.org/en/option.html#series-scatter.data}{series}}
}
\value{
A list for \emph{dataset.source} or \emph{series.data}. The latter does not include column names.
}
\description{
Get an EchartsJS dataset from a data.frame
}
