% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.dendro}
\alias{ec.dendro}
\title{Dendrogram}
\usage{
ec.dendro(hc)
}
\arguments{
\item{hc}{Result of \code{\link[stats]{hclust}} function.}
}
\value{
A tree structure, see format in \href{https://echarts.apache.org/en/option.html#series-tree.data}{tree data}.
}
\description{
Build data for Hierarchical Clustering dendrogram - a tree chart
}
\examples{
hc <- hclust(dist(USArrests), "complete")
p <- ec.init(preset=FALSE) 
p$x$opts$series <- list(list( 
  type= 'tree', orient= 'TB', roam= TRUE, initialTreeDepth= -1,
  data= ec.dendro(hc), 
  # layout= 'radial',
  # symbolSize= ec.clmn(-1, scale= 0.33), 
  #    exclude added labels like 'p99', leaving only the originals
  label= list(formatter= htmlwidgets::JS(
    "function(n) { out= /p\\\\d+/.test(n.name) ? '' : n.name; return out;}"))
))
p

}
