% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{ec.examples}
\alias{ec.examples}
\title{Code Examples}
\usage{
ec.examples()
}
\value{
No return value, used only for help
}
\description{
Learn by example - copy/paste code from Examples below.\cr
This code collection is to demonstrate various concepts of
data preparation, conversion, grouping,
parameter setting, visual fine-tuning,
custom rendering, plugins attachment,
Shiny plots & interactions through Shiny proxy.\cr
}
\seealso{
\href{https://helgasoft.github.io/echarty/}{website} has many more examples
}
