% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.fromJson}
\alias{ec.fromJson}
\title{Convert JSON string to chart}
\usage{
ec.fromJson(txt)
}
\arguments{
\item{txt}{JSON character string, url, or file, see \code{\link[jsonlite]{fromJSON}}}
}
\value{
An \code{echarty} widget.
}
\description{
Convert JSON string to chart
}
\details{
\code{txt} should contain the full list of options required to build a chart.
It is subsequently passed to EchartsJS function \href{https://echarts.apache.org/en/api.html#echartsInstance.setOption}{setOption}.
}
