% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.global}
\alias{ec.global}
\title{Global theme}
\usage{
ec.global(theme = NULL, font = NULL, urltiles = NULL)
}
\arguments{
\item{theme}{Name of theme file (without extension), from folder /inst/themes}

\item{font}{Font family name}

\item{urltiles}{Tiles URL template for leaflet maps}
}
\value{
none
}
\description{
Set a global theme, font and/or tile URL
}
\details{
To get these values in code use \code{\link[base]{getOption}}. Revert back to default by setting them to NULL.
}
\examples{
ec.global(theme = 'dark')
cars \%>\% ec.init()
ec.global(theme = NULL)
cars \%>\% ec.init()

}
