% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.inspect}
\alias{ec.inspect}
\title{Convert chart to JSON}
\usage{
ec.inspect(e, json = TRUE, ...)
}
\arguments{
\item{e}{An \code{echarty} widget as returned by \link{ec.init}}

\item{json}{Whether to return a JSON, or a \code{list}, default TRUE}

\item{...}{Additional options to pass to \code{\link[jsonlite]{toJSON}}}
}
\value{
A JSON string if \code{json} is \code{TRUE} and
a \code{list} otherwise.
}
\description{
Convert chart to JSON
}
\note{
Must be passed as last option.
}
\examples{
# extract JSON
json <- cars \%>\% ec.init() \%>\% ec.inspect()
json

# get from JSON and modify plot
ec.fromJson(json) \%>\% ec.theme('macarons')

}
