% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.js2r}
\alias{ec.js2r}
\title{JS-to-R Translator Assistant}
\usage{
ec.js2r()
}
\value{
none
}
\description{
Translate Javascript data objects to R
}
\details{
Learn from Javascript examples of \href{https://echarts.apache.org/examples/en/}{ECharts}. This Shiny application helps translate JSON-like JavaScript data structures to R lists.\cr
Additional information is available by clicking the \emph{Info} button inside.
}
