% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.snip}
\alias{ec.snip}
\title{Options list shortcut}
\usage{
ec.snip(wt)
}
\arguments{
\item{wt}{A widget to be converted to option list \cr
OR an option list to plot}
}
\description{
Utility to improve readability and typing speed
}
\details{
On initialization, pipe \emph{ec.snip} after \link{ec.init}, \cr
or set for the entire R session with \code{options('echarty.short'=TRUE)}.
}
\examples{
p <- cars |> ec.init() |> ec.snip()
p$dataZoom <- list(start=70)   # instead of p$x$opts$dataZoom
p$legend  <- list(zz='')       # instead of p$x$opts$tooltip
ec.snip(p)                     # instead of just p

}
