% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.theme}
\alias{ec.theme}
\title{Themes}
\usage{
ec.theme(e, name, code = NULL)
}
\arguments{
\item{e}{An \code{echarty} widget as returned by \link{ec.init}}

\item{name}{Name of existing theme file (without extension), or name of custom theme defined in \code{code}.}

\item{code}{Custom theme code in JSON format, default NULL.}
}
\value{
An \code{echarty} widget.
}
\description{
Apply a pre-built or custom coded theme to a chart
}
\details{
Just a few themes are included in folder \code{inst/themes}. The entire collection could be found \href{https://github.com/apache/echarts/tree/master/theme}{here} and copied if needed.
}
\examples{
mtcars \%>\% ec.init() \%>\% ec.theme('dark-mushroom')
cars \%>\% ec.init() \%>\% ec.theme('mine', code=
  '{"color": ["green","#eeaa33"], 
    "backgroundColor": "lemonchiffon"}')

}
