% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.vismap}
\alias{ec.vismap}
\title{Build EchartsJS visualMap list from data}
\usage{
ec.vismap(
  df = NULL,
  calculable = TRUE,
  type = c("continuous", "piecewise"),
  ...
)
}
\arguments{
\item{df}{Chart data in data.frame format, required}

\item{calculable}{Show handles or not, default TRUE}

\item{type}{One of 'continuous'(default) or 'piecewise'}

\item{...}{Any other \href{https://echarts.apache.org/en/option.html#visualMap}{options} to pass}
}
\value{
A list to set in parameter \code{visualMap}
}
\description{
Build EchartsJS visualMap list from data
}
