% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty-package.R
\docType{package}
\name{echarty-package}
\alias{echarty}
\alias{echarty-package}
\title{echarty: Minimal R/Shiny Interface to JavaScript Library 'ECharts'}
\description{
The goal is to deliver the full functionality of 'ECharts' with minimal overhead. 'ECharts' is based on data structures and 'echarty' users build R lists for these same data structures. One to three 'echarty' commands are usually sufficient to produce any chart.
}
\section{Package Conventions}{

echarty is "pipe-friendly", supports both \%>\% and |> \cr
echarty functions have three prefixes to help with auto-completion: \itemize{
\item \emph{ec.} for general functions
\item \emph{ecr.} for rendering functions
\item \emph{ecs.} for Shiny functions
}
}

\section{Global Options}{

Options are set with R command \href{https://www.rdocumentation.org/packages/base/versions/3.6.2/topics/options}{options}.\cr
echarty uses the following options: \cr \itemize{
\item echarty.theme = name of theme file (without extension), from folder /inst/themes
\item echarty.font = font family name
\item echarty.urltiles = tiles URL template for leaflet maps
\item echarty.short = a boolean flag, see \link{ec.snip}
}
}

\examples{
# basic scatter chart from a data.frame, using presets
cars |> ec.init()

# set/get global options
options('echarty.short'=TRUE)  # set
getOption('echarty.short')     # get
options('echarty.short'=NULL)  # remove

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/helgasoft/echarty}
  \item Report bugs at \url{https://github.com/helgasoft/echarty/issues/}
}

}
\author{
\strong{Maintainer}: Larry Helgason \email{larry@helgasoft.com} [copyright holder]

Authors:
\itemize{
  \item John Coene \email{jcoenep@gmail.com} [copyright holder]
}

}
\keyword{internal}
