% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ecs.proxy}
\alias{ecs.proxy}
\title{Shiny: Create a proxy}
\usage{
ecs.proxy(id)
}
\arguments{
\item{id}{Target chart id from the Shiny UI.}
}
\value{
A proxy object to update the chart.
}
\description{
Create a proxy for an existing chart in Shiny UI. It allows to
add, merge, delete elements to a chart without reloading it.
}
\seealso{
\link{ecs.exec} for example.
}
