% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
A value of the lhs (left-hand-side) object prepared for rhs (right-hand-side) function/call.
}
\description{
Imported from magrittr
}
\keyword{internal}
