% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.fromJson}
\alias{ec.fromJson}
\title{JSON to chart}
\usage{
ec.fromJson(txt, ...)
}
\arguments{
\item{txt}{JSON character string, url, or file, see \link[jsonlite]{fromJSON}}

\item{...}{Any attributes to pass to internal \link{ec.init}}
}
\value{
An \emph{echarty} widget.
}
\description{
Convert JSON string to chart
}
\details{
\emph{txt} could be either a list of options (x$opts) to pass to ECharts function \href{https://echarts.apache.org/en/api.html#echartsInstance.setOption}{setOption},\cr
or an entire \emph{htmlwidget} generated thru \code{ec.inspect(target='full')}.
}
\examples{
txt <- '{
   "xAxis": { "data": ["Mon", "Tue", "Wed"]},
   "yAxis": { },
   "series": { "type": "line", "data": [150, 230, 224] } }'
ec.fromJson(txt)

}
