% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.js2r}
\alias{ec.js2r}
\title{Translator Assistant}
\usage{
ec.js2r()
}
\value{
none
}
\description{
Translate Javascript data objects to R
}
\details{
To learn by examples in Javascript from \href{https://echarts.apache.org/examples/en/}{Echarts}
}
