% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.layout}
\alias{ec.layout}
\title{Charts layout}
\usage{
ec.layout(plots, rows = NULL, cols = NULL, width = "xs", title = NULL)
}
\arguments{
\item{plots}{A list of charts}

\item{rows}{Number of rows}

\item{cols}{Number of columns}

\item{width}{Width of columns, one of xs, md, lg}

\item{title}{Title for the set}
}
\value{
A container \code{\link[htmltools]{div}} in rmarkdown, otherwise \code{\link[htmltools]{browsable}}
}
\description{
Set multiple charts in rows/columns format
}
\examples{

if (interactive()) {
  p1 <- cars \%>\% ec.init()
  p2 <- cars \%>\% ec.init() \%>\% ec.theme('dark')
  ec.layout(list(p1,p2), cols=2 )
}

}
