% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.snip}
\alias{ec.snip}
\title{Options list shortcut}
\usage{
ec.snip(wt)
}
\arguments{
\item{wt}{A widget to be converted to option list \cr
OR an option list to a plot}
}
\description{
Utility to improve readability and typing speed
}
\details{
On initialization, add \emph{ec.snip} at the end of the \link{ec.init} pipe, \cr
or set for the entire R session with \code{options('echarty.short'=TRUE)}.\cr
Note: ec.theme, ecr.ebars, ec.inspect will not work with the session setting.
}
\examples{
p <- cars |> ec.init() |> ec.snip()
p$dataZoom <- list(start=70)   # instead of p$x$opts$dataZoom
p$legend  <- list(show=TRUE)   # instead of p$x$opts$tooltip
ec.snip(p)                     # instead of just p

}
