% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.upd}
\alias{ec.upd}
\title{Update option lists}
\usage{
ec.upd(wt, ...)
}
\arguments{
\item{wt}{An echarty widget}

\item{...}{A commands expression to update option lists}
}
\description{
Improve readability by chaining commands
}
\details{
Replaces syntax\cr
\verb{   }p <- ec.init(...)\cr
\verb{   }p$x$opts$series <- ...\cr
with\cr
\verb{   }ec.init(...) |> \verb{   } # set or preset chart params\cr
\verb{   }ec.upd(\{series <- ...\}) # update params thru R commands
}
\examples{
Orange |> dplyr::group_by(Tree) |> ec.init() |>
ec.upd({ 
  series <- lapply(series, function(x) {
    x$symbolSize= 10; x$encode= list(x='age', y='circumference'); x } )
})
}
