% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ec.util}
\alias{ec.util}
\title{Utility functions}
\usage{
ec.util(..., type = "sf.series", verbose = FALSE)
}
\arguments{
\item{...}{Optional parameters for \href{https://echarts.apache.org/en/option.html#series-scatter.type}{scatter}(points) or lines(polylines) charts.\cr}

\item{type}{Type of utility - 'sf.series'(default) or 'sf.bbox' \cr\itemize{
\item \emph{sf.series} returns a list of chart series\cr
\itemize{
\item required parameter \emph{df} - a result from \link[sf]{st_read}\cr
\item optional parameter \emph{nid} - column name for name-id used in tooltips\cr
\item \emph{sf.bbox} returns JavaScript code to position a map inside a bounding box from \link[sf]{st_bbox}\cr
}
}}

\item{verbose}{Print extra info in console, default is FALSE.}
}
\description{
currently related only to GIS shapefiles read by library sf
}
\details{
These utilities are experimental. Goal is to build map series from a shapefile.
Limitations:\cr no sfc_MULTIPOINT or sfc_POLYGON types supported yet,\cr
no polygons with holes, polygons can have only a name in tooltip,\cr
assumes Geodetic CRS is WGS 84.
}
\examples{
if (interactive()) {
  library(sf)
  fname <- system.file("shape/nc.shp", package="sf")
  nc <- st_read(fname)
  p <- ec.init(load= c('leaflet', 'custom'),  # load custom for polygons
         js= ec.util(type= 'sf.bbox', bbox= st_bbox(nc)),
         series= ec.util(df= nc, nid= 'NAME', itemStyle= list(opacity= 0.3)),
         tooltip= list(show= TRUE, formatter= '{a}')
  )
  p
}
}
