% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ecr.band}
\alias{ecr.band}
\title{Area band}
\usage{
ecr.band(df = NULL, lower = NULL, upper = NULL, type = "polygon", ...)
}
\arguments{
\item{df}{A data.frame with lower and upper numerical columns and first column with X coordinates.}

\item{lower}{The column name of band's lower boundary (string).}

\item{upper}{The column name of band's upper boundary (string).}

\item{type}{Type of rendering
\itemize{
\item 'polygon' - by drawing a polygon as polyline from upper/lower points (default)
\item 'stack' - by two \href{https://echarts.apache.org/en/option.html#series-line.stack}{stacked lines}
}}

\item{...}{More attributes for \href{https://echarts.apache.org/en/option.html#series-line.type}{serie}}
}
\value{
A list of \strong{one serie} when type='polygon', or list of \strong{two series} when type='stack'
}
\description{
A 'custom' serie with lower and upper boundaries
}
\details{
\itemize{
\item type='polygon': coordinates of the two boundaries are chained into one polygon. Tooltips do not show upper band values.\cr
\emph{xAxis type} could be 'category' or 'value'.\cr
Set fill color with attribute \emph{color}.\cr
\item type='stack': two \emph{stacked} lines are drawn, the lower with customizable areaStyle.\cr
\emph{xAxis type} should be 'category'! Tooltips could include upper band values.\cr
Set fill color with attribute \emph{areaStyle$color}.\cr
}
Optional parameter \emph{name}, if given, will show up in legend. Legend merges all series with same name into one item.
}
\examples{
df <- airquality |> dplyr::mutate(
    lwr= round(Temp-Wind*2),
    upr= round(Temp+Wind*2),
      x= paste0(Month,'-',Day) ) |>
  dplyr::relocate(x, Temp)
bands <- ecr.band(df, 'lwr', 'upr', type='stack',
                  name= 'stak', areaStyle= list(opacity=0.4))
df |> ec.init( load= 'custom',
   legend= list(show= TRUE),
   dataZoom= list(type= 'slider'),
   toolbox= list(feature= list(dataZoom= list(show= TRUE))),
   xAxis= list(type= 'category', boundaryGap= FALSE),
   series= list(
     list(type='line', color='blue', name='line'),
     bands[[1]], bands[[2]]
   ),
   tooltip= list( trigger= 'axis',
     formatter= ec.clmn(
        'high <b>\%@</b><br>line <b>\%@</b><br>low <b>\%@</b>',
                  3.3, 1.2, 2.2)
   )  # 3.3= upper_serie_index +.+ index_of_column_inside
)

}
