% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ecs.exec}
\alias{ecs.exec}
\title{Shiny: Execute a proxy command}
\usage{
ecs.exec(proxy, cmd = "p_merge")
}
\arguments{
\item{proxy}{A \link{ecs.proxy} object}

\item{cmd}{Name of command, default is \emph{p_merge}\cr
The proxy commands are:\cr
\emph{p_update} - add new series and axes\cr
\emph{p_merge} - modify or add series features like style,marks,etc.\cr
\emph{p_replace} - replace entire chart \cr
\emph{p_del_serie} - delete a serie by index or name\cr
\emph{p_del_marks} - delete marks of a serie\cr
\emph{p_append_data} - add data to existing series\cr
\emph{p_dispatch} - send action commands, see \href{https://echarts.apache.org/en/api.html#echartsInstance.dispatchAction}{documentation}}
}
\value{
A proxy object to update the chart.
}
\description{
Once chart changes had been made, they need to be sent back to the widget for display
}
\examples{
if (interactive()) {
   demo(eshiny, package='echarty')
}

}
\seealso{
\link{ecs.proxy}, \link{ecs.render}, \link{ecs.output} \cr
Read about event handling in \link{-- Introduction --}, or from \href{https://github.com/helgasoft/echarty/blob/main/demo/examples.R}{examples}.
}
