% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echarty.R
\name{ecs.render}
\alias{ecs.render}
\title{Shiny: Plot command to render chart}
\usage{
ecs.render(wt, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{wt}{An \code{echarty} widget to generate the chart.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression? default FALSE.}
}
\value{
An output or render function that enables the use of the widget within Shiny applications.
}
\description{
This is the initial rendering of a chart in the UI.
}
\seealso{
\link{ecs.exec} for example, \code{\link[htmlwidgets]{shinyRenderWidget}} for return value.
}
